:recycle: refactor: refactor prob with clean code, and typesafe (#4)
