+++ src/util/modelUtil.ts
